优化IM出现参数错误的问题

+++ app/Events/WorkerStartEvent.php
@@ -1,6 +1,8 @@
 <?php
+
 namespace App\Events;
 
+use Cache;
 use DB;
 use Hhxsv5\LaravelS\Swoole\Events\WorkerStartInterface;
 use Swoole\Http\Server;
@@ -14,7 +16,9 @@ class WorkerStartEvent implements WorkerStartInterface
 
     public function handle(Server $server, $workerId)
     {
-        // TODO: Implement handle() method.
-        DB::table('ws')->delete();
+        if (isset($server->startMsecTime) && Cache::get("swooleServerStartMsecTime") != $server->startMsecTime) {
+            Cache::forever("swooleServerStartMsecTime", $server->startMsecTime);
+            DB::table('ws')->delete();
+        }
     }
 }
